removed comparison test for obscollectionholders.
Those tests were meant to prevent multiple propagation in case there are no change. However keeping the changes in memory would actually require deep copy and be could still fail in case the internal structure refer to external data (eg int pointers), because in case those external data were changed, the hashcode of the holding structure would not change and thus would not propagate a change. the only solution is to propagate a change every time.